feat(buffer): Supporting a lock-free and thread-safe RingBuffer to address the issue of memory reuse
pref code style
