+++ pkg/math/math.go
@@ -8,3 +8,10 @@ func Max(a, b int) int {
 
 	return a
 }
+
+func Min(a, b int) int {
+	if a < b {
+		return a
+	}
+	return b
+}
