+++ pkg/math/math_test.go
@@ -13,3 +13,11 @@ func TestMax(t *testing.T) {
 	num2 := Max(2, 3)
 	assert.Equal(t, 3, num2)
 }
+
+func TestMin(t *testing.T) {
+	num := Min(1, 0)
+	assert.Equal(t, 0, num)
+
+	num2 := Min(2, 3)
+	assert.Equal(t, 2, num2)
+}
